Fix floating point operations (#22)
One notable change is Eq for Value is removed. The instance is not used in the
interpreter and should not be used in the future, because of the Float/Double
variants. A new valEq function is implemented to compare test outputs, which
compares floats bitwise.

New Show implementation for Value shows bit patterns of floats, to be used when
debugging.