Fix trapping in start function when allocating a module (#25)
According to the spec, a trap in a start function does not cause module
initialization to fail (spec doesn't explicitly say that it's OK to trap
in a start function, but it also doesn't say "fail", and the last test
in "linking.wast" suggests that initialization should not fail).

Previously, when start traps, changes in memories or tables would not be
reverted, but the module would be unavailable. This is not correct
behavior, but if it was, the code would still be buggy, because if the
failing module has `elem` sections the table would have references to
functions in a non-existent module.

This patch implements trapping in start according to the spec: the
module now becomes available when start traps.